update weight to decimal

+++ shipment/models.py
@@ -8,13 +8,13 @@ from shipment.choices import SHIPMENT_TYPE
 class ShippingFee(SoftDeleteModel):
     id = models.AutoField(primary_key=True)
     location = models.ForeignKey(State, on_delete=models.DO_NOTHING)
-    weight_start = models.IntegerField()
-    weight_end = models.IntegerField()
+    weight_start = models.DecimalField(max_digits=8, decimal_places=2)
+    weight_end = models.DecimalField(max_digits=8, decimal_places=2)
     ship_fee = models.DecimalField(max_digits=10, decimal_places=2)
     sub_fee = models.DecimalField(
         max_digits=10, decimal_places=2, blank=True, null=True
     )
-    sub_weight = models.IntegerField(blank=True, null=True)
+    sub_weight = models.DecimalField(max_digits=8, decimal_places=2,blank=True, null=True)
 
     class Meta:
         db_table = "shipping_fee"
